refactor: 인증정보에 대한 예외 조건 추가 - 인증 정보가 없을 시 AuthenticationException 발생 - SubwayExceptionHandler에 AuthenticationException 처리 메소드 추가
