+++ src/main/java/nextstep/exception/AuthenticationException.java
@@ -1,8 +1,14 @@
-package nextstep.member;
+package nextstep.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthenticationException extends RuntimeException {
+	public AuthenticationException() {
+	}
+
+	public AuthenticationException(String message) {
+		super(message);
+	}
 }
